// src/api/model/auditLog.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Inventory API
 * Inventory & Stock Management REST API — Repository + Service pattern, domain events, HMAC device webhooks, audit trail.
 * OpenAPI spec version: 1.0.0
 */
import type { AuditLogProperties } from './auditLogProperties';

export interface AuditLog {
  id?: number;
  log_name?: string;
  description?: string;
  /** @nullable */
  subject_type?: string | null;
  /** @nullable */
  subject_id?: number | null;
  /** @nullable */
  causer_type?: string | null;
  /** @nullable */
  causer_id?: number | null;
  /** @nullable */
  properties?: AuditLogProperties;
  created_at?: string;
}